organizeFiles
