+++ src/app/app-routing.module.ts
@@ -1,14 +1,19 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { DashboardComponent } from './pages/hero/dashboard/dashboard.component';
-import { HeroDetailComponent } from './pages/hero/hero-detail/hero-detail.component';
-import { HeroesComponent } from './pages/heroes/heroes.component';
+import { LayoutComponent } from './layout/layout/layout.component';
 
 const routes: Routes = [
-  { path: '', redirectTo: '/dashboard', pathMatch: 'full' },
-  { path: 'heroes', component: HeroesComponent },
-  { path: 'detail/:id', component: HeroDetailComponent },
-  { path: 'dashboard', component: DashboardComponent },
+  {
+    path: '',
+    component: LayoutComponent,
+    children: [
+      {
+        path: '',
+        loadChildren: () =>
+          import('./pages/hero/hero.module').then((m) => m.HeroModule),
+      },
+    ],
+  },
 ];
 //of 將物件轉成observable
 @NgModule({
