--- src/app/app.component.css
@@ -1,25 +1,0 @@
-
-.stock{
-  margin-right: 1rem;
-}
-/* AppComponent's private CSS styles */
-h1 {
-  margin-bottom: 0;
-}
-nav a {
-  padding: 1rem;
-  text-decoration: none;
-  margin-top: 10px;
-  display: inline-block;
-  background-color: #e8e8e8;
-  color: #3d3d3d;
-  border-radius: 4px;
-}
-
-nav a:hover {
-  color: white;
-  background-color: #3a14e2;
-}
-nav a.active {
-  background-color: black;
-}
