+++ src/app/app.component.html
@@ -1,12 +1,2 @@
-<h1>{{ title }}</h1>
-
-<nav>
-  <a routerLink="/heroes" class="stock">Heroes</a>
-  <a routerLink="/dashboard">dashboard</a>
-</nav>
-<!-- <ng-container *ngFor="let num of arrayA$ | async">
-  <div>{{ num }}</div>
-</ng-container> -->
 <router-outlet></router-outlet>
 
-<app-messages></app-messages>
