+++ src/app/app.module.ts
@@ -4,22 +4,12 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { DataPipe } from './data.pipe';
-import { DashboardComponent } from './pages/hero/dashboard/dashboard.component';
-import { HeroDetailComponent } from './pages/hero/hero-detail/hero-detail.component';
-import { HeroesComponent } from './pages/heroes/heroes.component';
-import { MessagesComponent } from './pages/messages/messages.component';
+import { LayoutComponent } from './layout/layout/layout.component';
+import { PipesModule } from './share/pipes/pipes.module';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    HeroesComponent,
-    HeroDetailComponent,
-    MessagesComponent,
-    DataPipe,
-    DashboardComponent,
-  ],
-  imports: [BrowserModule, FormsModule, AppRoutingModule],
+  declarations: [AppComponent  ,LayoutComponent],
+  imports: [BrowserModule, FormsModule, AppRoutingModule, PipesModule],
   providers: [],
   bootstrap: [AppComponent],
 })
