+++ src/app/core/mock/mock-heroes.ts
@@ -1,4 +1,5 @@
-import { Hero } from './hero';
+import { Hero } from "../model/hero";
+
 
 export const HEROES: Hero[] = [
   { id: 12, name: 'Dr. Nice' },
