+++ src/app/core/service/message.service.ts
@@ -7,10 +7,12 @@ export class MessageService {
   messages: string[] = [];
   add(message: string) {
     this.messages.push(message);
+    localStorage.setItem('info', JSON.stringify(this.messages));
   }
 
   clear() {
     this.messages = [];
+    localStorage.removeItem('info');
   }
 
   constructor() { }
