+++ src/app/layout/layout/layout.component.html
@@ -6,3 +6,4 @@
 </nav>
 
 <router-outlet></router-outlet>
+
