+++ src/app/pages/hero/hero-detail/hero-detail.component.html
@@ -16,3 +16,4 @@
   </div>
   <button type="button" (click)="goBack()">go back</button>
 </div>
+<app-messages></app-messages>
