+++ src/app/pages/hero/hero.module.ts
@@ -1,0 +1,23 @@
+import { LayoutComponent } from './../../layout/layout/layout.component';
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+
+
+import { DashboardComponent } from './dashboard/dashboard.component';
+import { HeroDetailComponent } from './hero-detail/hero-detail.component';
+import { HeroRoutingModule } from './hero-routing.module';
+import { HeroesComponent } from './heroes/heroes.component';
+import { MessagesComponent } from './messages/messages.component';
+
+@NgModule({
+  declarations: [
+    HeroesComponent,
+    HeroDetailComponent,
+    DashboardComponent,
+    MessagesComponent,
+  
+  ],
+  imports: [CommonModule, HeroRoutingModule, FormsModule],
+})
+export class HeroModule {}
