+++ src/app/pages/hero/heroes/heroes.component.html
@@ -32,3 +32,4 @@
   </li>
 </ul> -->
 <!-- -------------------------- -->
+<app-messages></app-messages>
