+++ src/app/share/pipes/data/data.pipe.ts
