+++ src/app/share/pipes/pipes.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { DataPipe } from './data/data.pipe';
+
+
+
+@NgModule({
+  declarations: [DataPipe],
+  imports: [
+    CommonModule
+  ]
+})
+export class PipesModule { }
